Fixed bug that prevented collection of latency metrics because the consul node name was not valid
On consul integration version 2.0.0, the following error is logged:

[ERR] Error collecting latency metrics for Agent '192.168.40.233:8301': could not find node for agent

The problem is that the entity name is a combination of the consul
address and port number, while the actual consul node name is not. Its
usually the FQDN but it does not have to be. Its extremely unlike to be
an IP and its not going to have a port number in it.

To fix this problem, the code for finding a node and returning its
api.Coordinate struct was modified to get the consul name from the
entity's ID Attributes.

I also added go.mod to enable module support so that people who are
runing in that mode can easily work and submit changes to this repo.